actions/requirements-actions.txt
rasa-sdk==3.6.18
psycopg2-binary==2.9.7
requests==2.31.0
openai==0.28.1